+++ src/main.rs
@@ -4,6 +4,7 @@ use crate::settings::RunCommand;
 use crate::settings::Settings;
 
 use std::error::Error;
+use std::process::Command;
 
 fn main() -> Result<(), Box<dyn Error>> {
     // The flow of the program is build around the idea of creating a main settings structure from
@@ -14,7 +15,7 @@ fn main() -> Result<(), Box<dyn Error>> {
     // structure.  The last step would be to actually execute the command and finish up the final
     // work.
 
-    let argument_options = Settings::new_from_cmdline(None)?;
+    let argument_options = Settings::new_from_cmdline(None);
 
     // If option `open-config` is set, then open the file and exit program.
     if argument_options.open_config()? {
@@ -29,35 +30,57 @@ fn main() -> Result<(), Box<dyn Error>> {
 
     let mut app_settings = Settings::new();
     // Overwrite fields in app_settings only, if new fields are Some().
-    app_settings.update_from(user_config)?;
-    app_settings.update_from(stdin_games)?;
-    app_settings.update_from(argument_options)?;
+    app_settings.update_from(user_config);
+    app_settings.update_from(stdin_games);
+    app_settings.update_from(argument_options);
 
-    // Without a game entry there is no point in running the program.
-    if !app_settings.is_game_available() {
-        return Err("A path to game is required.".into());
-    }
-
-    let mut defaults = Settings::new_from_defaults()?;
+    let mut defaults = Settings::new_from_defaults();
     if !app_settings.is_libretro_path_available() {
         // Extract keys and values from `retroarch.cfg` only if the path to `libretro` installation
         // directory in `RetroArch` is unknown.
         let raconfig = Settings::new_from_retroarch_config(
             app_settings.get_retroarch_config(),
         )?;
-        defaults.update_from(raconfig)?;
+        defaults.update_from(raconfig);
     }
     // Overwrite only those keys in `app_settings`, which their values are currently `None`.
-    app_settings.update_defaults_from(defaults)?;
+    app_settings.update_defaults_from(defaults);
 
-    // Build the final commandline for RetroArch and execute it, if permitted.
-    let mut run: RunCommand = app_settings.build_command()?;
-    if app_settings.there_can_only_be_one() {
-        eprintln!("retroarch process already running. There Can Be Only One!");
+    if app_settings.is_game_available() {
+        let mut run: RunCommand = app_settings.build_command()?;
+
+        if app_settings.there_can_only_be_one() {
+            eprintln!(
+                "retroarch process already running. There Can Be Only One!"
+            );
+        } else {
+            run.output = app_settings.run(&mut run.cmdline);
+        }
+        if app_settings.is_list_cores() {
+            for core in app_settings.find_core_match(&run.libretro) {
+                println!("{core}");
+            }
+        }
+        if app_settings.is_which_command() {
+            print_cmdline(&run.cmdline);
+        } else {
+            app_settings.print_which(run.game);
+        }
+    } else if app_settings.is_list_cores() {
+        app_settings.print_cores();
     } else {
-        run.output = app_settings.run(&mut run.cmdline);
+        return Err("A path to game is required.".into());
     }
-    app_settings.print_which(run.game);
 
     Ok(())
+}
+
+// Prints program name and each commandline arguments exactly the same as it is used to run
+// RetroArch.
+fn print_cmdline(command: &Command) {
+    print!("{:?}", command.get_program());
+    for arg in command.get_args() {
+        print!(" {:?}", arg);
+    }
+    println!();
 }